remove staging table
simplify code

fix test

remove old staging tables
